The "finished" n-body problem code!
So this is the nearly-final version of the n-body problem. I wrote it way back in august. It's nearly final because I wanna add a Barnes-Hut algorithm to this to make it like a cherry on top but I'm currently struggling with its implementation. Even without it, the program has many inconsistencies such as energy not conserving or angular momentum not being defined properly, so it's still pretty much unfinished. But there is already an animation that gives somewhat real result. So I've decided to post it. If you want to check out the actual unfinished version with all the "trash" in it, you can find it in the text file I've also added.

The code itself contains a class of a gravitational system with a function that simulates the Newtonian gravitational interaction and a main part of the code that can vary depending on initial conditions. Here, the initial conditions are: all masses are almost equal with small deviations, they are located in the grid nodes of the square of the minimum side that can cover them all, and at the start they don't move. My animation is kinda laggy when you pick a big number, so I've set N = 10 by default to not crash your computer accidentaly :)

Also, I've added some GIF's that are basically some experiments. However, the mass distribution is different cause those GIF's are also from august. The gif named gravity_simulation is just a regular gravity interaction animation; gravity_test1 is a different world where F ∝ -1/r and gravity_test2 is if gravitational force was not attractive, but repulsive(you can just think about it as if it was an electric force and the bodies are, for example, electrons with q ∝ m); black_hole.gif is an apocaliptic result of adding an object with an enormous mass into gravity_simulation.gif(not really a black hole because Einstein said no).